Move part of bootstrap into blockchain module
Some blockchain-related bootstrap operations were handled in the
network module and that impaired a bit the separation of tasks
netween those two modules.
Therefore, that part of the bootstrap process has been moved
to the blockchain module
